refactor: move shared request routing into the dapp-request domain
